+++ src/app/shared/component/table/directives/table-column.directive.ts
@@ -1,0 +1,10 @@
+import { Directive, Input, TemplateRef } from '@angular/core';
+
+@Directive({
+  selector: '[tableColumn]'
+})
+export class TableColumnDirective {
+  @Input('header') header!: string;
+
+  constructor(public templateRef: TemplateRef<any>) {}
+}